fix: properly copy Dates
Fix #164
Firestore has native support for timestamps, which show up in JS as `Date` objects. We need a special case in `extractRefs` because recursing into a `Date` and copying its fields doesn't produce a proper copy. Instead we should pass it through the same way we do with more fundamental types, like strings or numbers.