feat(core): add reset option to rtdb
BREAKING CHANGE: When unbinding in RTDB, the data is, by default reset
to `null` for items bound as objects, and to an empty array `[]` to
items bound as an array. This option option can be set to `false` to
keep the last value instead of resetting it, or to a function returning
a value to customize the new value.